refactor(webapp): extract shared MonthStepper
MonthPagination and MonthSelector duplicated the arrow-stepper logic,
earliest-month tooltip, and month math. Extract a shared MonthStepper
(header / compact variants); the two become thin wrappers.
